refactor: [step5] 인터페이스 주입받도록 변경
