+++ README.md
@@ -53,3 +53,4 @@
 - [X] ...... 객체 내부에서 처리하도록 메시지 전달
 - [X] ...... 객체를 비교하도록 테스트 코드 수정
 - [X] ...... 내부 구현 변경, 사용하지 않는 내부변수 수정
+- [X] ...... 인터페이스 주입받도록 변경
