+++ src/main/java/racing/service/NumberGenerator.java
@@ -1,0 +1,5 @@
+package racing.service;
+
+public interface NumberGenerator {
+    public Integer getNumber();
+}
