+++ src/main/java/racing/service/RacingService.java
@@ -5,16 +5,15 @@ import racing.domain.Count;
 import racing.domain.RacingCars;
 import racing.domain.RacingHistory;
 import racing.view.ResultView;
-import util.RandomNumberGenerator;
 
 import java.util.Set;
 
 public class RacingService {
 
     public static void race(String carNames, Count tryCount) throws CarLocationException {
-        RandomNumberGenerator randomNumberGenerator = new RandomNumberGenerator();
+        NumberGenerator numberGenerator = new RandomNumberGenerator();
         RacingCars racingCars = new RacingCars(carNames);
-        Set<RacingHistory> racingHistory = racingCars.startRace(tryCount, randomNumberGenerator);
+        Set<RacingHistory> racingHistory = racingCars.startRace(tryCount, numberGenerator);
         ResultView.showRaceResult(racingHistory);
         ResultView.showRaceWinner(racingCars);
     }
