+++ src/main/java/racing/service/RandomNumberGenerator.java
@@ -1,12 +1,12 @@
-package util;
+package racing.service;
 
 import java.util.Random;
 
-public class RandomNumberGenerator {
+public class RandomNumberGenerator implements NumberGenerator{
 
     private static final int RANDOM_INT_MAX_LIMIT = 10;
     private static final Random random = new Random();
-    public static Integer getRandomNumber() {
+    public Integer getNumber() {
         return random.nextInt(RANDOM_INT_MAX_LIMIT);
     }
 }
