+++ src/test/java/racing/RacingServiceTest.java
@@ -7,8 +7,9 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import racing.domain.Car;
 import racing.domain.Location;
+import racing.service.NumberGenerator;
+import racing.service.RandomNumberGenerator;
 import util.RacingValidator;
-import util.RandomNumberGenerator;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -42,10 +43,11 @@ public class RacingServiceTest {
     @Test
     @DisplayName("자동차 우승을 가리기 위해 최대 이동거리를 구하는 테스트를 수행한다.")
     void 리스트들중_최대_숫자구하기_테스트() {
+        NumberGenerator randomNumberGenerator = new RandomNumberGenerator();
         List<Integer> numList = new ArrayList<>();
         int maxValue = Integer.MIN_VALUE;
         for (int i = 0; i < 4; i++) {
-            int randomNumber = RandomNumberGenerator.getRandomNumber();
+            int randomNumber = randomNumberGenerator.getNumber();
             maxValue = Math.max(maxValue, randomNumber);
             numList.add(randomNumber);
         }
