+++ src/test/java/util/RandomNumberGeneratorTest.java
@@ -2,6 +2,8 @@ package util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.service.NumberGenerator;
+import racing.service.RandomNumberGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -10,7 +12,8 @@ public class RandomNumberGeneratorTest {
     @Test
     @DisplayName("0에서 9 사이의 랜덤 숫자를 얻는다.")
     void 랜덤_숫자_얻기_테스트() {
-        Integer randomNumber = RandomNumberGenerator.getRandomNumber();
+        NumberGenerator randomNumberGenerator = new RandomNumberGenerator();
+        Integer randomNumber = randomNumberGenerator.getNumber();
         assertThat(isNumberInRange(randomNumber)).isTrue();
     }
